+++ R/storage.R
@@ -1,12 +1,14 @@
 #' Retrieve Data from MongoDB
 #'
-#' This function connects to a MongoDB database and retrieves all documents from a specified collection.
+#' This function connects to a MongoDB database and retrieves all documents from a specified collection,
+#' maintaining the original column order if available.
 #'
 #' @param connection_string A character string specifying the MongoDB connection URL. Default is NULL.
 #' @param collection_name A character string specifying the name of the collection to query. Default is NULL.
 #' @param db_name A character string specifying the name of the database. Default is NULL.
 #'
-#' @return A data frame containing all documents from the specified collection.
+#' @return A data frame containing all documents from the specified collection, with columns ordered
+#'         as they were when the data was originally pushed to MongoDB.
 #'
 #' @keywords storage
 #'
@@ -22,21 +24,41 @@
 #'
 #' @export
 mdb_collection_pull <- function(connection_string = NULL, collection_name = NULL, db_name = NULL) {
-  data <- mongolite::mongo(collection = collection_name, db = db_name, url = connection_string)
-  data$find()
+  # Connect to the MongoDB collection
+  collection <- mongolite::mongo(collection = collection_name, db = db_name, url = connection_string)
+
+  # Retrieve the order document
+  order_doc <- collection$find(query = '{"type": "column_order"}')
+
+  if (nrow(order_doc) == 0) {
+    warning("Column order information not found. The order of variables may not match the original.")
+    return(collection$find(query = '{"type": {"$ne": "column_order"}}'))
+  }
+
+  # Get the original column order
+  original_order <- order_doc$columns[[1]]
+
+  # Retrieve the data, excluding the order document
+  data <- collection$find(query = '{"type": {"$ne": "column_order"}}')
+
+  # Reorder the columns
+  data <- data[, original_order]
+
+  return(data)
 }
 
 #' Upload Data to MongoDB and Overwrite Existing Content
 #'
 #' This function connects to a MongoDB database, removes all existing documents
-#' from a specified collection, and then inserts new data.
+#' from a specified collection, and then inserts new data. It also stores the
+#' original column order to maintain data structure consistency.
 #'
 #' @param data A data frame containing the data to be uploaded.
 #' @param connection_string A character string specifying the MongoDB connection URL.
 #' @param collection_name A character string specifying the name of the collection.
 #' @param db_name A character string specifying the name of the database.
 #'
-#' @return The number of documents inserted.
+#' @return The number of data documents inserted into the collection (excluding the order document).
 #'
 #' @keywords storage
 #'
@@ -53,6 +75,15 @@ mdb_collection_pull <- function(connection_string = NULL, collection_name = NULL
 #'
 #' @export
 mdb_collection_push <- function(data = NULL, connection_string = NULL, collection_name = NULL, db_name = NULL) {
+  # Store the original column names
+  original_columns <- names(data)
+
+  # Create a special document to store the original column order
+  order_doc <- list(
+    type = "column_order",
+    columns = original_columns
+  )
+
   # Connect to the MongoDB collection
   collection <- mongolite::mongo(
     collection = collection_name,
@@ -63,11 +94,14 @@ mdb_collection_push <- function(data = NULL, connection_string = NULL, collectio
   # Remove all existing documents in the collection
   collection$remove("{}")
 
-  # Insert the new data
-  result <- collection$insert(data)
+  # Insert the order document first
+  collection$insert(order_doc)
 
-  # Return the number of documents inserted
-  return(result)
+  # Insert the new data
+  collection$insert(data)
+
+  # Return the number of documents in the collection (excluding the order document)
+  return(collection$count() - 1)
 }
 
 #' Get metadata tables